Use new handleSkillCommand method for all occurances of old server skill method plus some warning cleanup.
